Move stryker4s.extension.mutationtype to stryker4s.mutation
