+++ modules/core/src/main/scala/stryker4s/mutation/EqualityOperator.scala
@@ -1,4 +1,4 @@
-package stryker4s.extension.mutationtype
+package stryker4s.mutation
 
 import scala.meta.Term
