+++ modules/core/src/main/scala/stryker4s/mutation/MutationTypes.scala
@@ -1,4 +1,4 @@
-package stryker4s.extension.mutationtype
+package stryker4s.mutation
 
 import cats.syntax.option.*
 import stryker4s.extension.TreeExtensions.IsEqualExtension
@@ -9,6 +9,12 @@ import scala.meta.*
   */
 sealed trait Mutation[+T <: Tree] {
   def mutationName: String
+  /**
+    * Full name including the package.
+    * 
+    * E.g. stryker4s.mutation.BooleanLiteral
+    */
+  def fullName: String
 }
 
 object Mutation {
@@ -44,24 +50,29 @@ trait SubstitutionMutation[T <: Tree] extends Mutation[T] with NoInvalidPlacemen
 
 trait EqualityOperator extends SubstitutionMutation[Term.Name] {
   override val mutationName: String = classOf[EqualityOperator].getSimpleName
+  override def fullName: String = classOf[EqualityOperator].getName
 }
 
 trait BooleanLiteral extends SubstitutionMutation[Lit.Boolean] {
   override val mutationName: String = classOf[BooleanLiteral].getSimpleName
+  override def fullName: String = classOf[BooleanLiteral].getName
 }
 
 trait ConditionalExpression extends SubstitutionMutation[Lit.Boolean] {
   override val mutationName: String = classOf[ConditionalExpression].getSimpleName
+  override def fullName: String = classOf[ConditionalExpression].getName
 }
 
 trait LogicalOperator extends SubstitutionMutation[Term.Name] {
   override val mutationName: String = classOf[LogicalOperator].getSimpleName
+  override def fullName: String = classOf[LogicalOperator].getName
 }
 
 /** T &lt;: Term because it can be either a `Lit.String` or `Term.Interpolation`
   */
 trait StringLiteral[T <: Term] extends SubstitutionMutation[T] {
   override val mutationName: String = classOf[StringLiteral[?]].getSimpleName
+  override def fullName: String = classOf[StringLiteral[?]].getName
 }
 
 /** Base trait for method mutation
@@ -73,6 +84,7 @@ trait MethodExpression extends Mutation[Term] {
   protected val methodName: String
 
   override val mutationName: String = classOf[MethodExpression].getSimpleName
+  override def fullName: String = classOf[MethodExpression].getName
 
   def apply(f: String => Term): Term = f(methodName)
