+++ modules/core/src/test/scala/stryker4s/mutants/tree/MutantInstrumenterTest.scala
@@ -4,9 +4,9 @@ import cats.data.{NonEmptyList, NonEmptyVector}
 import fs2.io.file.Path
 import stryker4s.extension.TreeExtensions.*
 import stryker4s.extension.exception.UnableToBuildPatternMatchException
-import stryker4s.extension.mutationtype.{ConditionalTrue, GreaterThan, Mutation, True}
 import stryker4s.model.*
 import stryker4s.mutants.applymutants.ActiveMutationContext
+import stryker4s.mutation.{ConditionalTrue, GreaterThan, Mutation, True}
 import stryker4s.testkit.{LogMatchers, Stryker4sSuite}
 import stryker4s.testutil.TestData
