+++ modules/core/src/test/scala/stryker4s/mutation/MutationTypesTest.scala
@@ -1,4 +1,4 @@
-package stryker4s.extension.mutationtype
+package stryker4s.mutation
 
 import stryker4s.extension.ImplicitMutationConversion.mutationToTree
 import stryker4s.testkit.Stryker4sSuite
